Implemented Service Layer for Cards
